feat: restructuring of routes
