+++ src/routes/(protected)/search/+page.svelte
