+++ src/routes/(public)/+layout.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import Toolbar from '$lib/components/Global/Toolbar.svelte'
+</script>
+
+<Toolbar />
+<slot />
