+++ src/routes/(public)/landing/simple-hiring/simpleHiring.svelte
