+++ src/routes/(public)/logout/+page.server.ts
